[fix] CSV syntax for resource type group seeds
Whitespaces between the columns can lead to unexpected
results. For example, if the row is:

```
"amorphous_mandalore_1", "inorganic"
```

Then the resourceGroup column will literally be saved as:

```
 "inorganic"
```

Note there's a leading space and the quotation marks
are also persisted.

The fix is to remove the leading white space, such as:

```
"amorphous_mandalore_1","inorganic"
```